chore: refactored unit and contract test to accomodate refactored retrieveTotalVisitsByStatusAndOrgId UC and GW
